support request page (#223)
* support request page

* fixes

* more fixes

* Update .eslintrc.js

* finalizes the supreqpage

* Update src/pages/index.tsx


* Update src/pages/index.tsx
