+++ .env.sample
@@ -8,6 +8,7 @@ DISCORD_BOT_TOKEN="***********************************************************"
 DISCORD_BOT_CHANNEL_IDS="0123456789012345678,0123456789012345678" # Comma seperated channel ids
 
 JUDGE_SECRET="secret"
+SUPPORT_SECRET="secret"
 ADMIN_SECRET="something secretive"
 CHALLENGE_URL="https://your-sponsor-challenge.com"
