+++ src/api/middleware/requireAuth.ts
@@ -2,9 +2,10 @@ import { Handler } from 'express';
 import logger from '../../logger';
 
 const adminSecret = process.env.ADMIN_SECRET;
+const supportSecret = process.env.SUPPORT_SECRET;
 
 export const requireAuth = (redirect = false): Handler => (req, res, next): void => {
-  if (req.signedCookies?.authed === 'yes' || req.headers.authorization === adminSecret) {
+  if (req.signedCookies?.authed === 'yes' || req.headers.authorization === adminSecret || req.headers.authorization === supportSecret) {
     next();
   } else if (redirect) {
     res.redirect('/login');
